requirements.txt
black==23.9.0
click==8.1.7
pytest==7.4.2
pytest-cov==4.1.0
pylint==2.17.5
nbval==0.10.0
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
httpx==0.26.0
requests==2.31.0
colorlog==6.8.0
oauth2client==4.1.3
gspread==5.12.4
gspread-dataframe==3.3.1
pandas==2.1.4